[osd_cli] Improved command line interface on xinerama setups
When the whole screen consists of a multitude of monitors with different
resolutions the cli is not always unvisible. The cli always started on
the left most monitor. If the height of this monitor is smaller than one
of the other ones, the cli was presented in an unvisible area of the
whole screen.

The new method determines the actual monitor on which the focused client
is present. This monitor is used to display the cli.