Bugfix for loosing focus for mainly java applications
These applications use a so called FocusProxy window to handle the focus
for the whole java app. The real UI is running in as the parent of the
FocusProxy window. Now if such an app has the focus and the desktop is
switched to another and back again the focus is kept. Moreover
maximizing by a keyboard shortcuts has also not worked before.

Conflicts:
	debian/changelog